Use a pointer to a state so we can eventually get things like indentation to automatically always be right.  This doesn't do much now except bind everything about the output into one place.
There's an uncomfortable mix of parameter passing and use of a global
pointer which will have to get resolved.  Essentially we sometimes need
to in effect "push the state" onto a stack and make temporary changes
(such as extra indents). Since make isn't multithreaded we don't truly
need to parameterise everything but I am so prejudiced against using
globals that it's hard to finally make that decision.

There's also a beginning to the code for generating a list of
directories.  Again this is difficult - I'd like most of the code not to
be in the make source files so it's easy to merge but make's structure
doesn't really allow this since data structures are hidden within a
module almost like a class. I can export the ones I need and use them
but then of course it perverts that encapsulation and I actually have to
change the make source more to get it to happen than if I just tacked on
a function to the end of dir.c.

So that is also an experiment to revisit later.